Create individual files for each detector type.
