+++ larpandora/LArPandoraInterface/LArPandoraDetectorType.h
@@ -1,3 +1,6 @@
+#ifndef LAR_PANDORA_DETECTOR_TYPE_H
+#define LAR_PANDORA_DETECTOR_TYPE_H 1
+
 #include "larpandora/LArPandoraInterface/LArPandoraGeometry.h"
 
 #include "larcore/Geometry/Geometry.h"
@@ -17,23 +20,5 @@ namespace lar_pandora{
     };
 
     LArPandoraDetectorType *GetDetectorType();
-
-    /*
-    class DUNEFarDetHD : public LArPandoraDetectorType {
-    public:
-        bool ShouldSwitchUV() const override;
-    };
-    */
-
-    class DUNEFarDetVDThreeView : public LArPandoraDetectorType {
-    public:
-        float TargetViewU() const override {return 0.f; };
-        float TargetViewV() const override {return 0.f; };
-        float TargetViewW() const override {return 0.f; };
-        float WirePitch(const geo::View_t view) const override {return 0.f; };
-        float WireAngle(const geo::View_t view, const int tpc, const int cstat) const override {return 0.f; };
-        bool ShouldSwitchUV(const unsigned int tpc, const unsigned int cstat) const override {return false; };
-        void LoadDetectorGaps(LArDetectorGapList& listOfGaps) override {return; }; 
-    };
-
 }
+#endif
